[Feature][Connector-V2][Paimon] Rename package name
